fix: improve display and optimize known issues

+++ jds/game-allstop.sh
@@ -6,6 +6,8 @@
 #
 # https://github.com/honeok/archive/raw/master/jds/game-allstop.sh
 
+version='v0.0.3 (2024.12.31)'
+
 yellow='\033[93m'
 red='\033[31m'
 green='\033[92m'
@@ -14,37 +16,55 @@ _yellow() { echo -e "${yellow}$*${white}"; }
 _red() { echo -e "${red}$*${white}"; }
 _green() { echo -e "${green}$*${white}"; }
 
+_info_msg() { echo -e "\033[48;5;220m\033[1m提示${white} $*"; }
+_err_msg() { echo -e "\033[41m\033[1m警告${white} $*"; }
+_suc_msg() { echo -e "\033[42m\033[1m成功${white} $*"; }
+
 server_range=$(find /data/ -maxdepth 1 -type d -name "server*" | sed 's:.*/::' | grep -E '^server[0-9]+$' | sed 's/server//' | sort -n)
 
 clear
-cd /data/tool || exit 1
+_yellow "当前脚本版本: $version"
+
+# 操作系统和权限校验
+[ "$(id -ru)" -ne "0" ] && _err_msg "$(_red '需要root用户才能运行！')" && exit 1
+os_info=$(grep ^ID= /etc/*release | awk -F'=' '{print $2}' | sed 's/"//g')
+[[ "$os_info" != "debian" && "$os_info" != "ubuntu" && "$os_info" != "centos" && "$os_info" != "rhel" && "$os_info" != "rocky" && "$os_info" != "almalinux" ]] && exit 0
+
+cd /data/tool || { _err_msg "$(_red '/data/tool路径错误')" && exit 1; }
 if pgrep -f processcontrol-allserver.sh >/dev/null 2>&1; then
     pkill -9 -f processcontrol-allserver.sh
     [ -f "control.txt" ] && : > control.txt
     [ -f "dump.txt" ] && : > dump.txt
-    _green "processcontrol进程已终止文件已清空"
+    _suc_msg "$(_green 'processcontrol进程已终止文件已清空')"
 else
-    _red "processcontrol进程未运行无需终止"
+    _info_msg "$(_red 'processcontrol进程未运行无需终止')"
 fi
 
-cd /data/server/login || exit 1
+cd /data/server/login || { _err_msg "$(_red 'login服务器路径错误')" && exit 1; }
 ./server.sh stop
-_green "login服务器已停止"
+_suc_msg "$(_green 'login服务器已停止')"
 
-cd /data/server/gate || exit 1
+cd /data/server/gate || { _err_msg "$(_red 'gate服务器路径错误')" && exit 1; }
 ./server.sh stop
 sleep 60
-_green "gate服务器已停止"
+_suc_msg "$(_green 'gate服务器已停止')"
 
 for server_num in $server_range; do
     (
-        cd "/data/server$server_num/game" || return
+        if [ ! -d "/data/server$server_num/game" ]; then
+            _err_msg "$(_red "server${server_num}不存在，子进程已退出")"
+            exit 1 # 子进程中的退出，防止继续执行
+        fi
+
+        cd "/data/server$server_num/game" 2>/dev/null || { _err_msg "$(_red "server${server_num}路径错误")" && exit 1; }
+
         _yellow "正在处理server$server_num"
         ./server.sh flush
         sleep 60
         ./server.sh stop
     ) &
 done
-# 等待所有并行操作完成
-wait
-_green "所有Game服务器已完成flush和stop操作"
+
+wait # 等待并行任务
+
+_suc_msg "$(_green '所有game服务器已完成flush和stop操作')"
